requirements.txt
Flask==0.12.2
Flask-Migrate==2.2.1
Flask-Script==2.0.6
SQLAlchemy==1.1.11
Flask-SQLAlchemy==2.3.2
psycopg2-binary==2.7.5
alembic==0.9.9
python-dotenv==0.7.1
